+++ src/badge/constants.ts
@@ -1,18 +1,19 @@
+import { Theme } from '@emotion/react';
 import { BadgeColor, BadgeColorValues } from './interfaces';
-import { ecosystemTheme } from '../theme-provider/constants';
 
-export const BadgeColorMap: Record<keyof typeof BadgeColor, BadgeColorValues> =
-	{
-		grey: {
-			base: ecosystemTheme.colors.grey.base,
-			hover: ecosystemTheme.colors.grey[300],
-		},
-		primary: {
-			base: ecosystemTheme.colors.primary.base,
-			hover: ecosystemTheme.colors.secondary[300],
-		},
-		contrast: {
-			base: ecosystemTheme.colors.contrast.base,
-			hover: ecosystemTheme.colors.contrast[300],
-		},
-	};
+export const BadgeColorMap = (
+	theme: Theme
+): Record<keyof typeof BadgeColor, BadgeColorValues> => ({
+	grey: {
+		base: theme.colors.grey.base,
+		hover: theme.colors.grey[300],
+	},
+	primary: {
+		base: theme.colors.primary.base,
+		hover: theme.colors.secondary[300],
+	},
+	contrast: {
+		base: theme.colors.contrast.base,
+		hover: theme.colors.contrast[300],
+	},
+});