+++ src/table/index.ts
@@ -1,0 +1,3 @@
+export * from './interfaces';
+export * from './table';
+export * from './table.style';