+++ src/theme-provider/index.tsx
@@ -1,12 +1,2 @@
-import React from 'react';
-import { ThemeProvider as NativeThemeProvider, Theme } from '@emotion/react';
-import { ecosystemTheme } from './constants';
-
-export const ThemeProvider: React.FC<{
-	children: React.ReactNode;
-	theme?: Theme;
-}> = ({ children, theme }) => (
-	<NativeThemeProvider theme={theme ?? ecosystemTheme}>
-		{children}
-	</NativeThemeProvider>
-);
+export * from './constants';
+export * from './theme-provider';